feat: introduce purchase page with a specific route
