+++ packages/profile/src/components/app.tsx
@@ -25,8 +25,10 @@ import { Slot } from "#components/slot";
 import { useMemo } from "react";
 import { Socials } from "./socials";
 import { Leaderboard } from "./leaderboard";
+import { CollectionPurchase } from "./inventory/collection/collection-purchase";
 
 export function App() {
+  console.log({ url: window.location.href });
   const [searchParams] = useSearchParams();
   const social = useMemo(() => {
     return searchParams.get("social");
@@ -56,9 +58,11 @@ export function App() {
                 <Route path="token/:tokenId" element={<CollectionAsset />}>
                   <Route path="send" element={<SendCollection />} />
                   <Route path="list" element={<CollectionListing />} />
+                  <Route path="purchase" element={<CollectionPurchase />} />
                 </Route>
                 <Route path="send" element={<SendCollection />} />
                 <Route path="list" element={<CollectionListing />} />
+                <Route path="purchase" element={<CollectionPurchase />} />
               </Route>
               <Route path="collectible/:address" element={<Collectible />}>
                 <Route path="token/:tokenId" element={<CollectibleAsset />}>
