+++ packages/profile/src/components/inventory/collection/collection-asset.tsx
@@ -27,26 +27,15 @@ import {
   Token,
   Spinner,
   Thumbnail,
-  ThumbnailCollectible,
-  Select,
-  TokenSelectHeader,
-  SelectContent,
-  SelectItem,
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-  InfoIcon,
 } from "@cartridge/ui";
 
-import { cn, useCountervalue } from "@cartridge/ui/utils";
+import { cn } from "@cartridge/ui/utils";
 import {
   AllowArray,
   cairo,
   Call,
   CallData,
   constants,
-  getChecksumAddress,
   TransactionExecutionStatus,
   TransactionFinalityStatus,
 } from "starknet";
@@ -84,18 +73,9 @@ export function CollectionAsset() {
   const { theme } = useTheme();
   const { editions } = useArcade();
   const { tokens } = useTokens();
-  const {
-    isListed,
-    provider,
-    selfOrders,
-    order,
-    removeOrder,
-    marketplaceFee,
-    royaltyFee,
-    setAmount,
-  } = useMarketplace();
+  const { isListed, provider, selfOrders, order, removeOrder, setAmount } =
+    useMarketplace();
   const [loading, setLoading] = useState(false);
-  const [validated, setValidated] = useState(false);
 
   const edition: EditionModel | undefined = useMemo(() => {
     return Object.values(editions).find(
@@ -163,138 +143,9 @@ export function CollectionAsset() {
     return Number(order?.price) / Math.pow(10, token?.metadata.decimals || 0);
   }, [order, token]);
 
-  const { total, fees } = useMemo(() => {
-    const price = Number(order?.price);
-    const formattedMarketplaceFee =
-      marketplaceFee / Math.pow(10, token?.metadata.decimals || 0);
-    const formattedRoyaltyFee =
-      Number(royaltyFee) / Math.pow(10, token?.metadata.decimals || 0);
-    const total = formattedMarketplaceFee + formattedRoyaltyFee;
-    const fees = [
-      {
-        label: "Marketplace Fee",
-        amount: `${formattedMarketplaceFee} ${token?.metadata.symbol}`,
-        percentage: `${(price ? (marketplaceFee / price) * 100 : 0).toFixed(2)}%`,
-      },
-      {
-        label: "Creator Royalties",
-        amount: `${formattedRoyaltyFee} ${token?.metadata.symbol}`,
-        percentage: `${(price ? (Number(royaltyFee) / price) * 100 : 0).toFixed(2)}%`,
-      },
-    ];
-    return { total, fees };
-  }, [marketplaceFee, royaltyFee, token]);
-
-  const { countervalues } = useCountervalue(
-    {
-      tokens: [
-        {
-          balance: `${amount}`,
-          address: token?.metadata.address || "",
-        },
-      ],
-    },
-    { enabled: !!amount && !!token },
-  );
-
-  const price: number = useMemo(() => {
-    if (!countervalues) return 0;
-    return countervalues[0]?.current.value || 0;
-  }, [countervalues]);
-
-  const purchaseData = useMemo(() => {
-    if (!assets || !collection || !price || !token || !order || !amount)
-      return {
-        assets: [],
-        currency: { name: "", image: "", price: 0, value: "" },
-      };
-    const tokens = assets.map((asset) => ({
-      name: asset.name,
-      image: asset.imageUrl || collection.imageUrl || placeholder,
-      collection: collection.name,
-    }));
-    const currency = {
-      name: token.metadata.symbol,
-      image: token.metadata.image || "",
-      price: amount,
-      value: price ? `~$${price.toFixed(2)}` : "",
-    };
-    return { assets: tokens, currency };
-  }, [assets, collection, price, token, order, amount, placeholder]);
-
   const handleBack = useCallback(() => {
     navigate(`..?${searchParams.toString()}`);
   }, [navigate, searchParams]);
-
-  const handlePurchase = useCallback(async () => {
-    if (!contractAddress || !asset || !isListed || !order || isOwner) return;
-    setLoading(true);
-    try {
-      const marketplaceAddress: string = provider.manifest.contracts.find(
-        (c: { tag: string }) => c.tag?.includes("Marketplace"),
-      )?.address;
-      const calls: AllowArray<Call> = [
-        {
-          contractAddress: getChecksumAddress(order.currency),
-          entrypoint: "approve",
-          calldata: CallData.compile({
-            spender: marketplaceAddress,
-            amount: cairo.uint256(order.price),
-          }),
-        },
-        {
-          contractAddress: marketplaceAddress,
-          entrypoint: "execute",
-          calldata: CallData.compile({
-            orderId: order.id,
-            collection: contractAddress,
-            tokenId: cairo.uint256(asset.tokenId),
-            assetId: cairo.uint256(asset.tokenId),
-            quantity: 0, // 0 for ERC721
-            royalties: true,
-          }),
-        },
-      ];
-      const res = await parent.openExecute(
-        Array.isArray(calls) ? calls : [calls],
-        chainId,
-      );
-      if (res?.transactionHash) {
-        await mainProvider.waitForTransaction(res.transactionHash, {
-          retryInterval: 1000,
-          successStates: [
-            TransactionExecutionStatus.SUCCEEDED,
-            TransactionFinalityStatus.ACCEPTED_ON_L2,
-          ],
-        });
-      }
-      if (res) {
-        toast.success(`Asset purchased successfully`);
-      }
-      setValidated(false);
-      // Removing the order optimistically
-      removeOrder(order);
-    } catch (error) {
-      console.error(error);
-      toast.error(`Failed to purchase asset(s)`);
-    } finally {
-      setLoading(false);
-    }
-  }, [
-    contractAddress,
-    asset,
-    isListed,
-    chainId,
-    parent,
-    provider,
-    mainProvider,
-    order,
-    isOwner,
-    removeOrder,
-    navigate,
-    setValidated,
-    searchParams,
-  ]);
 
   const handleUnlist = useCallback(async () => {
     if (!contractAddress || !asset || !isListed || !order || !isOwner) return;
@@ -387,17 +238,15 @@ export function CollectionAsset() {
 
   if (
     location.pathname.includes("/send") ||
-    location.pathname.includes("/list")
+    location.pathname.includes("/list") ||
+    location.pathname.includes("/purchase")
   ) {
     return <Outlet />;
   }
 
   return (
     <LayoutContainer>
-      <LayoutHeader
-        className="hidden"
-        onBack={validated ? () => setValidated(false) : handleBack}
-      />
+      <LayoutHeader className="hidden" onBack={handleBack} />
 
       {status === "loading" || !collection || !asset ? (
         <LoadingState />
@@ -405,93 +254,94 @@ export function CollectionAsset() {
         <EmptyState />
       ) : (
         <>
-          {validated ? (
-            <PurchaseConfirmation {...purchaseData} />
-          ) : (
-            <LayoutContent className="p-6 pb-0 flex flex-col gap-6 overflow-hidden">
-              <CollectionHeader
-                image={edition?.properties.icon || theme?.icon}
-                title={title}
-                subtitle={collection.name}
-                expiration={
-                  isListed && selfOrders.length > 0
-                    ? selfOrders[0].expiration
-                    : undefined
-                }
-                listingCount={isListed ? selfOrders.length : undefined}
+          <LayoutContent
+            className={cn(
+              "p-6 flex flex-col gap-6 overflow-hidden",
+              isListed || (isOwner && "pb-0"),
+            )}
+          >
+            <CollectionHeader
+              image={edition?.properties.icon || theme?.icon}
+              title={title}
+              subtitle={collection.name}
+              expiration={
+                isListed && selfOrders.length > 0
+                  ? selfOrders[0].expiration
+                  : undefined
+              }
+              listingCount={isListed ? selfOrders.length : undefined}
+            />
+            <div
+              className="flex flex-col gap-6 overflow-scroll relative"
+              style={{ scrollbarWidth: "none" }}
+            >
+              <CollectiblePreview
+                image={asset.imageUrl || collection.imageUrl || placeholder}
+                size="lg"
+                className="w-full self-center mt-0.5"
               />
-              <div
-                className="flex flex-col gap-6 overflow-scroll relative"
-                style={{ scrollbarWidth: "none" }}
+              {!!order && (
+                <div className="absolute top-[-2px] right-2">
+                  <Price amount={amount} image={token?.metadata.image} />
+                </div>
+              )}
+              <CollectibleTabs
+                order={["details", "activity"]}
+                className={cn("pb-0", isListed || (isOwner && "pb-6"))}
               >
-                <CollectiblePreview
-                  image={asset.imageUrl || collection.imageUrl || placeholder}
-                  size="lg"
-                  className="w-full self-center mt-0.5"
-                />
-                {!!order && (
-                  <div className="absolute top-[-2px] right-2">
-                    <Price amount={amount} image={token?.metadata.image} />
-                  </div>
-                )}
-                <CollectibleTabs
-                  order={["details", "activity"]}
-                  className="pb-6"
+                <TabsContent
+                  className="m-0 p-0 flex flex-col gap-y-4"
+                  value="details"
                 >
-                  <TabsContent
-                    className="m-0 p-0 flex flex-col gap-y-4"
-                    value="details"
-                  >
-                    {properties.length > 0 && (
-                      <CollectibleProperties properties={properties} />
-                    )}
-                    <CollectibleDetails
-                      chainId={chainId as constants.StarknetChainId}
-                      address={collection.address}
-                      tokenId={asset.tokenId}
-                      standard={collection.type}
-                      owner={username}
-                    />
-                  </TabsContent>
-                  <TabsContent
-                    className="m-0 p-0 flex flex-col gap-px"
-                    value="activity"
-                  >
-                    {events.map((props: CardProps, index: number) => (
-                      <Link
-                        key={`${index}-${props.key}`}
-                        to={to(props.transactionHash)}
-                        target="_blank"
-                      >
-                        <TraceabilityCollectibleCard
-                          username={props.username}
-                          timestamp={props.timestamp}
-                          category={props.category}
-                          amount={props.amount}
-                          collectibleImage={
-                            asset.imageUrl || collection.imageUrl || placeholder
-                          }
-                          collectibleName={title || collection.name}
-                          currencyImage={props.currencyImage}
-                        />
-                      </Link>
-                    ))}
-                    <Button
-                      variant="secondary"
-                      className={cn(
-                        "text-foreground-300 hover:text-foreground-200 normal-case text-sm font-medium tracking-normal font-sans",
-                        cap >= data.length && "hidden",
-                      )}
-                      onClick={() => setCap((prev) => prev + OFFSET)}
+                  {properties.length > 0 && (
+                    <CollectibleProperties properties={properties} />
+                  )}
+                  <CollectibleDetails
+                    chainId={chainId as constants.StarknetChainId}
+                    address={collection.address}
+                    tokenId={asset.tokenId}
+                    standard={collection.type}
+                    owner={username}
+                  />
+                </TabsContent>
+                <TabsContent
+                  className="m-0 p-0 flex flex-col gap-px"
+                  value="activity"
+                >
+                  {events.map((props: CardProps, index: number) => (
+                    <Link
+                      key={`${index}-${props.key}`}
+                      to={to(props.transactionHash)}
+                      target="_blank"
                     >
-                      <PlusIcon variant="solid" size="xs" />
-                      See More
-                    </Button>
-                  </TabsContent>
-                </CollectibleTabs>
-              </div>
-            </LayoutContent>
-          )}
+                      <TraceabilityCollectibleCard
+                        username={props.username}
+                        timestamp={props.timestamp}
+                        category={props.category}
+                        amount={props.amount}
+                        collectibleImage={
+                          asset.imageUrl || collection.imageUrl || placeholder
+                        }
+                        collectibleName={title || collection.name}
+                        currencyImage={props.currencyImage}
+                      />
+                    </Link>
+                  ))}
+                  <Button
+                    variant="secondary"
+                    className={cn(
+                      "text-foreground-300 hover:text-foreground-200 normal-case text-sm font-medium tracking-normal font-sans",
+                      cap >= data.length && "hidden",
+                    )}
+                    onClick={() => setCap((prev) => prev + OFFSET)}
+                  >
+                    <PlusIcon variant="solid" size="xs" />
+                    See More
+                  </Button>
+                </TabsContent>
+              </CollectibleTabs>
+            </div>
+          </LayoutContent>
 
           <LayoutFooter
             className={cn(
@@ -499,39 +349,6 @@ export function CollectionAsset() {
               !isListed && !isOwner && "hidden",
             )}
           >
-            <div
-              className={cn(
-                "flex flex-col gap-3 w-full",
-                !validated && "hidden",
-              )}
-            >
-              <div className="w-full px-3 py-2.5 flex flex-col gap-1 bg-background-125 border border-background-200 rounded">
-                <div className="flex items-center justify-between text-xs text-foreground-400">
-                  <p>Cost</p>
-                  <p>{`${amount - total} ${purchaseData.currency.name}`}</p>
-                </div>
-                <div className="flex items-center justify-between text-xs text-foreground-400">
-                  <div className="flex gap-2  text-xs font-medium">
-                    Fees
-                    <FeesTooltip trigger={<InfoIcon size="xs" />} fees={fees} />
-                  </div>
-                  <p>{`${total} ${purchaseData.currency.name}`}</p>
-                </div>
-              </div>
-              <div className="flex gap-3 w-full">
-                <div className="w-full px-3 py-2.5 flex items-center justify-between bg-background-125 border border-background-200 rounded">
-                  <p className="text-sm font-medium text-foreground-400">
-                    Total
-                  </p>
-                  <p className="text-sm font-medium text-foreground-100">{`${purchaseData.currency.price * purchaseData.assets.length}`}</p>
-                </div>
-                <TokenSelect
-                  tokens={tokens}
-                  selected={token || tokens[0]}
-                  setSelected={() => {}}
-                />
-              </div>
-            </div>
             <div className="flex gap-3 w-full">
               <Button
                 variant="secondary"
@@ -539,7 +356,7 @@ export function CollectionAsset() {
                 onClick={handleUnlist}
                 className={cn(
                   "w-full gap-2 text-destructive-100",
-                  (validated || !isListed || !isOwner) && "hidden",
+                  (!isListed || !isOwner) && "hidden",
                 )}
               >
                 <TagIcon variant="solid" size="sm" />
@@ -548,7 +365,7 @@ export function CollectionAsset() {
               <Link
                 className={cn(
                   "flex items-center justify-center gap-x-4 w-full",
-                  (validated || isListed || !isOwner) && "hidden",
+                  (isListed || !isOwner) && "hidden",
                 )}
                 to={`list?${searchParams.toString()}`}
               >
@@ -557,21 +374,25 @@ export function CollectionAsset() {
                   List
                 </Button>
               </Link>
-              <Button
-                isLoading={loading}
-                variant="primary"
-                className={cn(
-                  "w-full gap-2",
-                  (validated || !isListed || isOwner) && "hidden",
-                )}
-                onClick={() => setValidated(true)}
-              >
-                Purchase
-              </Button>
               <Link
                 className={cn(
                   "flex items-center justify-center gap-x-4 w-full",
-                  (validated || !isOwner) && "hidden",
+                  (!isListed || isOwner) && "hidden",
+                )}
+                to={`purchase?${searchParams.toString()}`}
+              >
+                <Button
+                  isLoading={loading}
+                  variant="primary"
+                  className="w-full gap-2"
+                >
+                  Purchase
+                </Button>
+              </Link>
+              <Link
+                className={cn(
+                  "flex items-center justify-center gap-x-4 w-full",
+                  !isOwner && "hidden",
                 )}
                 to={`send?${searchParams.toString()}`}
               >
@@ -580,31 +401,6 @@ export function CollectionAsset() {
                   Send
                 </Button>
               </Link>
-              <div
-                className={cn(
-                  "w-full flex items-center gap-3",
-                  !validated && "hidden",
-                )}
-              >
-                <Button
-                  variant="secondary"
-                  type="button"
-                  className="w-1/3"
-                  onClick={() => navigate(`../..?${searchParams.toString()}`)}
-                >
-                  Cancel
-                </Button>
-                <Button
-                  type="submit"
-                  className="w-2/3"
-                  isLoading={loading}
-                  onClick={
-                    validated ? handlePurchase : () => setValidated(true)
-                  }
-                >
-                  {validated ? "Confirm" : "Review"}
-                </Button>
-              </div>
             </div>
           </LayoutFooter>
         </>
@@ -612,85 +408,6 @@ export function CollectionAsset() {
     </LayoutContainer>
   );
 }
-
-const PurchaseConfirmation = ({
-  assets,
-  currency,
-}: {
-  assets: {
-    name: string;
-    image: string;
-    collection: string;
-  }[];
-  currency: {
-    name: string;
-    image: string;
-    price: number;
-    value: string;
-  };
-}) => {
-  return (
-    <>
-      <LayoutContent className="p-6 pb-0 flex flex-col gap-4 overflow-hidden select-none">
-        <div className="h-10 flex items-center justify-start gap-3 select-none">
-          <Thumbnail
-            icon={
-              <TagIcon
-                variant="solid"
-                size="lg"
-                className="h-[30px] w-[30px]"
-              />
-            }
-            size="lg"
-          />
-          <p className="text-lg/[24px] font-semibold">Review Purchase</p>
-        </div>
-        <div
-          className="grow flex flex-col gap-px rounded overflow-y-scroll pb-6"
-          style={{ scrollbarWidth: "none" }}
-        >
-          <div className="h-10 flex items-center justify-between bg-background-200 px-3 py-2.5">
-            <p className="text-xs tracking-wider text-foreground-400 font-semibold">
-              Purchasing
-            </p>
-            <p
-              className={cn(
-                "px-1.5 py-0.5 text-xs bg-background-300 rounded-full font-medium text-foreground-300",
-                assets.length <= 1 && "hidden",
-              )}
-            >{`${assets.length} total`}</p>
-          </div>
-          {assets.map((asset, index) => (
-            <div
-              key={`${asset.name}-${index}`}
-              className="h-16 flex items-center justify-between bg-background-200 px-4 py-3 gap-3"
-            >
-              <ThumbnailCollectible
-                image={asset.image}
-                size="lg"
-                variant="light"
-                className="p-0"
-              />
-              <div className="flex flex-col gap-0.5 items-stretch grow overflow-hidden">
-                <div className="flex items-center gap-6 justify-between text-sm font-medium capitalize">
-                  <p>{asset.name}</p>
-                  <div className="flex items-center gap-1">
-                    <Thumbnail icon={currency.image} size="sm" />
-                    <p>{currency.price}</p>
-                  </div>
-                </div>
-                <div className="flex items-center gap-1 justify-between text-xs text-foreground-300">
-                  <p className="truncate">{asset.collection}</p>
-                  <p>{currency.value}</p>
-                </div>
-              </div>
-            </div>
-          ))}
-        </div>
-      </LayoutContent>
-    </>
-  );
-};
 
 const Price = ({ amount, image }: { amount?: number; image?: string }) => {
   const containerRef = useRef<HTMLDivElement>(null);
@@ -733,107 +450,6 @@ const Price = ({ amount, image }: { amount?: number; image?: string }) => {
         />
       </div>
     </div>
-  );
-};
-
-const TokenSelect = ({
-  tokens,
-  selected,
-  setSelected,
-}: {
-  tokens: Token[];
-  selected: Token;
-  setSelected: (token: Token) => void;
-}) => {
-  const onChangeToken = useCallback(
-    (address: string) => {
-      const token = tokens.find((token) => token.metadata.address === address);
-      if (token) {
-        setSelected(token);
-      }
-    },
-    [setSelected, tokens],
-  );
-
-  if (tokens.length === 0) return null;
-
-  return (
-    <Select
-      value={selected?.metadata.address}
-      onValueChange={onChangeToken}
-      defaultValue={selected?.metadata.address}
-      disabled // TODO: enable when swap is implemented
-    >
-      <TokenSelectHeader className="h-10 w-fit rounded flex gap-2 items-center p-2 bg-background-200 hover:bg-background-200 disabled:opacity-100" />
-      <SelectContent viewPortClassName="gap-0 bg-background-100 flex flex-col gap-px">
-        {tokens.map((token) => (
-          <SelectItem
-            key={token.metadata.address}
-            simplified
-            value={token.metadata.address}
-            data-active={token.metadata.address === selected?.metadata.address}
-            className="h-10 group bg-background-200 hover:bg-background-300 text-foreground-300 hover:text-foreground-100 cursor-pointer data-[active=true]:bg-background-200 data-[active=true]:hover:bg-background-300 data-[active=true]:text-foreground-100 rounded-none"
-          >
-            <div className="flex items-center gap-2">
-              {token.metadata.image ? (
-                <Thumbnail
-                  icon={token.metadata.image}
-                  rounded
-                  size="sm"
-                  variant="light"
-                  className="group-hover:bg-background-400"
-                />
-              ) : (
-                <div className="w-5 h-5 bg-gray-200 rounded-full flex-shrink-0" />
-              )}
-              <span className="font-medium text-sm">
-                {token.metadata.symbol}
-              </span>
-            </div>
-          </SelectItem>
-        ))}
-      </SelectContent>
-    </Select>
-  );
-};
-
-export const FeesTooltip = ({
-  trigger,
-  fees,
-}: {
-  trigger: React.ReactNode;
-  fees: {
-    label: string;
-    amount: string;
-    percentage: string;
-  }[];
-}) => {
-  return (
-    <TooltipProvider>
-      <Tooltip>
-        <TooltipTrigger>{trigger}</TooltipTrigger>
-        <TooltipContent
-          side="top"
-          align="start"
-          className="px-3 py-2 flex flex-col gap-2 bg-spacer-100 border border-background-150 text-foreground-400 min-w-[240px] select-none"
-        >
-          {fees.map((fee) => (
-            <div
-              key={fee.label}
-              className="flex flex-row justify-between text-foreground-300"
-            >
-              {fee.label}:{" "}
-              <div>
-                <span className="text-foreground-400 text-xs">
-                  ({fee.percentage})
-                </span>{" "}
-                {fee.amount}
-              </div>
-            </div>
-          ))}
-        </TooltipContent>
-      </Tooltip>
-    </TooltipProvider>
   );
 };
